[SDSAET-50] Rename, move and delete some classes concerning grouping
