+++ core/jobs/src/main/java/com/cognifide/aet/job/common/groupers/GroupingAlgorithm.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2013 Cognifide Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License. You may obtain a copy of the License at
+ * in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -13,13 +13,12 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.groupers.algorithm;
+package com.cognifide.aet.job.common.groupers;
 
-public class GroupingException extends Exception {
+import java.util.Collection;
+import java.util.Set;
 
-  private static final long serialVersionUID = -7006642565625349049L;
+public interface GroupingAlgorithm<T> {
 
-  public GroupingException(String message) {
-    super(message);
-  }
+  Set<Set<T>> group(Collection<T> elementsToGroup) throws GroupingException;
 }
