+++ core/jobs/src/main/java/com/cognifide/aet/job/common/groupers/GroupingException.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2013 Cognifide Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License. You may obtain a copy of the License at
+ * in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -13,12 +13,13 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.groupers.algorithm;
+package com.cognifide.aet.job.common.groupers;
 
-import java.util.Collection;
-import java.util.Set;
+public class GroupingException extends Exception {
 
-public interface GroupingAlgorithm<T> {
+  private static final long serialVersionUID = -7006642565625349049L;
 
-  Set<Set<T>> group(Collection<T> elementsToGroup) throws GroupingException;
+  public GroupingException(String message) {
+    super(message);
+  }
 }
